Work on blood fluid
This commit adds the water tag to the blood fluid, to get the water
physics etc. Also adds a couple of client-side fog events to change the
blood fog from blue (due to the water tag) to red. Also tweaks some of
the blood fluid attributes to be more slow-flowing.