Add blood biome
This commit adds a very bare-bone blood biome. Also, due to the fact
that biomes are registered before features (annoying), changes had to be
made regarding how features and configured features are initialized and
registered, to make sure that they are initialized and not null when the
biome needs them.