Special loot when fishing in blood
This commit changes the loot obtained from fishing in blood. Previously,
the same loot tables as vanilla water fishing was used. Now, with the
use of global loot modifiers, special blood fishing loot tables are
used. As of right now, only blood runes can be fished in blood, but this
is only temporary.